# swagger/ecomm.yaml
openapi: 3.0.3
info:
  title: Swagger ECOMM - OpenAPI 3.0
  description: API utilizada no curso Alura LevelUp | <NxtDev/> - PagoNxt
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:3000
...
paths:
...
  /api/admin/users:
    get:
      tags:
        - user
      summary: Get all users
      description: Retorna todos usuários.
      operationId: findUsers
      responses:
        "200":
          description: SUCESSO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users"
        "404":
          description: Nenhum usuário encontrado.
          content:
            application/json:
              example: "Not found"
    post:
      tags:
        - user
      summary: Create one user
      description: Cria um usuário.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createUser"
      responses:
        "201":
          description: Usuário criado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "404":
          description: Erro ao criar usuário.
          content:
            application/json:
              example: "Not found"
  /api/users/{id}:
    get:
      tags:
        - user
      summary: Get one user
      description: Retorna um usuário especifico.
      operationId: findUserById
      parameters:
        - name: id
          in: path
          description: Id do usuário.
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        "200":
          description: SUCESSO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "404":
          description: Nenhum usuário encontrado.
          content:
            application/json:
              example: "Not found"
  /api/admin/users/update/{id}:
    put:
      tags:
        - user
      summary: Update a user
      description: Atualiza um usuário.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: Id do usuário.
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createUser"
      responses:
        "200":
          description: Usuário atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "404":
          description: Usuário não encontrado.
          content:
            application/json:
              example: "Not found"
        "400":
          description: Requisição incorreta.
          content:
            application/json:
              example: "Bad Request"
  /api/admin/users/remove/{id}:
    delete:
      tags:
        - user
      summary: Remove a user
      description: Remove um usuário.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: Id do usuário.
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        "200":
          description: Usuário removido com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "404":
          description: Usuário não encontrado.
          content:
            application/json:
              example: "Not found"
        "400":
          description: Requisição incorreta.
          content:
            application/json:
              example: "Bad Request"
...
components:
  schemas:
...
    user:
      type: object
      properties:
        _id:
          oneOf:
            - type: string
            - type: integer
        username:
          type: string
        email:
          type: string
        senha:
          type: string
        cpf:
          type: string
        telefone:
          type: string
        dataCriacao:
          type: string
        endereco:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
    users:
      type: array
      items:
        $ref: "#/components/schemas/user"
    createUser:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        senha:
          type: string
        cpf:
          type: string
        telefone:
          type: string
        endereco:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
...